Struct libp2p_yamux::Config

pub struct Config { /* private fields */ }

The yamux configuration.

Implementations

impl Config

pub fn client() -> Self

Creates a new YamuxConfig in client mode, regardless of whether it will be used for an inbound or outbound upgrade.

pub fn server() -> Self

Creates a new YamuxConfig in server mode, regardless of whether it will be used for an inbound or outbound upgrade.

pub fn set_receive_window_size(&mut self, num_bytes: u32) -> &mut Self

Sets the size (in bytes) of the receive window per substream.

pub fn set_max_buffer_size(&mut self, num_bytes: usize) -> &mut Self

Sets the maximum size (in bytes) of the receive buffer per substream.

pub fn set_max_num_streams(&mut self, num_streams: usize) -> &mut Self

Sets the maximum number of concurrent substreams.

pub fn set_window_update_mode(&mut self, mode: WindowUpdateMode) -> &mut Self

Sets the window update mode that determines when the remote is given new credit for sending more data.

pub fn into_local(self) -> LocalConfig

Converts the config into a LocalConfig for use with upgrades of I/O streams that are ![Send].